+++ src/main/java/com/smart/tailor/entities/Material.java
@@ -35,7 +35,7 @@ public class Material extends AuditEntity implements Serializable {
     private String unit;
 
     @Column(name = "base_price")
-    private Double basePrice;
+    private Integer basePrice;
 
     private Boolean status;
 }
